added new RootNavigator, edit Confirmation Screen, App, navigator file names"
