+++ src/navigators/AuthNavigator.tsx
