+++ src/navigators/HomeNavigator.tsx
